// frontend/mrsexam-ui/src/api/generated/models/FormPandit.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FormPandit = {
    rollNo: number;
    sessionNo: number;
    name?: string | null;
    formCentre?: string | null;
    examCentre?: string | null;
    rollNo1?: number | null;
    sessionNo1?: number | null;
    marks1?: string | null;
    rollNo2?: number | null;
    sessionNo2?: number | null;
    marks2?: string | null;
    rollNo3?: number | null;
    sessionNo3?: number | null;
    marks3?: string | null;
    rollNo4?: number | null;
    sessionNo4?: number | null;
    marks4?: string | null;
    rollNo5?: number | null;
    sessionNo5?: number | null;
    marks5?: string | null;
    rollNo6?: number | null;
    sessionNo6?: number | null;
    marks6?: string | null;
    oralRollNo?: number | null;
    oralSessionNo?: number | null;
    oralMarks?: string | null;
    optMarks?: string | null;
    khand1TotMarks?: number | null;
    khand2TotMarks?: number | null;
    totMarks?: number | null;
    grade?: string | null;
    undFlag?: string | null;
    undClass?: string | null;
    specialRank?: string | null;
    khandAppearing?: string | null;
    bothKhandFlag?: string | null;
    othKhandRollNo?: number | null;
    concessionFlag?: string | null;
    resDeclFlag?: string | null;
    result?: string | null;
    add1?: string | null;
    add2?: string | null;
    add3?: string | null;
    dummyCentreCode?: number | null;
    userId?: string | null;
};